Read functions now explicitly use 64bit pointers

+++ fs-lib/src/file.rs
@@ -8,15 +8,15 @@ use anyhow::{Result, bail};
 use crate::buffer::BufferExtension;
 
 pub trait FileExtension {
-    fn read_bytes(&mut self, offset: usize, length: usize) -> Result<Vec<u8>>;
-    fn read_u8(&mut self, offset: usize) -> Result<u8>;
-    fn read_u16(&mut self, offset: usize) -> Result<u16>;
-    fn read_u32(&mut self, offset: usize) -> Result<u32>;
-    fn read_string(&mut self, offset: usize, length: usize) -> Result<String>;
+    fn read_bytes(&mut self, offset: u64, length: usize) -> Result<Vec<u8>>;
+    fn read_u8(&mut self, offset: u64) -> Result<u8>;
+    fn read_u16(&mut self, offset: u64) -> Result<u16>;
+    fn read_u32(&mut self, offset: u64) -> Result<u32>;
+    fn read_string(&mut self, offset: u64, length: usize) -> Result<String>;
 }
 
 impl FileExtension for File {
-    fn read_bytes(&mut self, offset: usize, length: usize) -> Result<Vec<u8>> {
+    fn read_bytes(&mut self, offset: u64, length: usize) -> Result<Vec<u8>> {
         let mut buffer: Vec<u8> = vec![0; length];
 
         self.seek(SeekFrom::Start(offset as u64))?;
@@ -33,25 +33,25 @@ impl FileExtension for File {
         Ok(buffer)
     }
 
-    fn read_u8(&mut self, offset: usize) -> Result<u8> {
+    fn read_u8(&mut self, offset: u64) -> Result<u8> {
         let buffer = self.read_bytes(offset, 1)?;
 
         Ok(buffer[0])
     }
 
-    fn read_u16(&mut self, offset: usize) -> Result<u16> {
+    fn read_u16(&mut self, offset: u64) -> Result<u16> {
         let buffer = self.read_bytes(offset, 2)?;
 
         Ok(buffer.read_u16(0))
     }
 
-    fn read_u32(&mut self, offset: usize) -> Result<u32> {
+    fn read_u32(&mut self, offset: u64) -> Result<u32> {
         let buffer = self.read_bytes(offset, 4)?;
 
         Ok(buffer.read_u32(0))
     }
 
-    fn read_string(&mut self, offset: usize, length: usize) -> Result<String> {
+    fn read_string(&mut self, offset: u64, length: usize) -> Result<String> {
         let buffer = self.read_bytes(offset, length)?;
 
         buffer.read_string(0, length)
